[test] RIGHT-TO-LEFT-MARK with IGNORE
RIGHT-TO-LEFT-MARK (U+200F or RLM) characters are
non-printing characters for setting the correct
right-to-left writing direction in mixed left-to-right
and right-to-left texts. On some operating systems or
in some settings, keyboard input puts RLM before
letters of right-to-left scripts.

Using IGNORE, it's possible to remove these characters.

The problem reported by Viktor Kovács.